02/requirements.txt
absl-py==2.3.0
altair==5.5.0
annotated-types==0.7.0
array-record==0.4.1
asgiref==3.8.1
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
chardet==3.0.4
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Django==4.2.22
dm-tree==0.1.8
etils==1.5.2
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.0
fsspec==2025.5.1
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
googletrans==4.0.0rc1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.9.0
h2==3.2.0
h5py==3.13.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
image==1.5.33
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kagglehub==0.3.12
keras==3.10.0
keras-core==0.1.7
keras-cv==0.9.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
ml_dtypes==0.5.1
mpld3==0.5.10
mpmath==1.3.0
namex==0.1.0
narwhals==1.41.0
networkx==3.2.1
numpy==2.0.2
oauthlib==3.2.2
opt_einsum==3.4.0
optree==0.16.0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
promise==2.3
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.8
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3986==1.5.0
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlparse==0.5.3
streamlit==1.45.1
sympy==1.13.1
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-datasets==4.9.3
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.1
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.17.1
termcolor==3.1.0
threadpoolctl==3.6.0
toml==0.10.2
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tornado==6.5.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
watchdog==6.0.0
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.22.0
